% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanPlacement.R
\name{clean_placement}
\alias{clean_placement}
\title{Clean placement data}
\usage{
clean_placement(data)
}
\arguments{
\item{data}{A data frame like placement}
}
\value{
A cleaned data frame
}
\description{
Clean placement data
}
