% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation_cgpa_placement}
\alias{correlation_cgpa_placement}
\title{Correlation between CGPA and placement}
\usage{
correlation_cgpa_placement(data)
}
\arguments{
\item{data}{Cleaned placement data}
}
\value{
Numeric correlation
}
\description{
Correlation between CGPA and placement
}
