% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation_internship_placement}
\alias{correlation_internship_placement}
\title{Correlation between internship and placement}
\usage{
correlation_internship_placement(data)
}
\arguments{
\item{data}{Cleaned placement data}
}
\value{
Numeric correlation
}
\description{
Correlation between internship and placement
}
