% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation_iq_placement}
\alias{correlation_iq_placement}
\title{Correlation between IQ and placement}
\usage{
correlation_iq_placement(data)
}
\arguments{
\item{data}{Cleaned placement data}
}
\value{
Numeric correlation
}
\description{
Correlation between IQ and placement
}
