% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Correlation matrix for key variables}
\usage{
correlation_matrix(data)
}
\arguments{
\item{data}{Cleaned placement data}
}
\value{
Correlation matrix
}
\description{
Correlation matrix for key variables
}
