% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ggplot.R
\name{plot_college_placement_rate_ggplot}
\alias{plot_college_placement_rate_ggplot}
\title{ggplot2: Placement rate by college}
\usage{
plot_college_placement_rate_ggplot(data)
}
\description{
ggplot2: Placement rate by college
}
