% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{plot_internship_counts_base}
\alias{plot_internship_counts_base}
\title{Base R: Internship completion counts}
\usage{
plot_internship_counts_base(data)
}
\description{
Base R: Internship completion counts
}
