% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ggplot.R
\name{plot_internship_counts_ggplot}
\alias{plot_internship_counts_ggplot}
\title{ggplot2: Internship completion counts}
\usage{
plot_internship_counts_ggplot(data)
}
\description{
ggplot2: Internship completion counts
}
