% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{plot_iq_group_placement_rate_base}
\alias{plot_iq_group_placement_rate_base}
\title{Base R: Placement rate by IQ group}
\usage{
plot_iq_group_placement_rate_base(data)
}
\description{
Base R: Placement rate by IQ group
}
