% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ggplot.R
\name{plot_iq_group_placement_rate_ggplot}
\alias{plot_iq_group_placement_rate_ggplot}
\title{ggplot2: Placement rate by IQ group}
\usage{
plot_iq_group_placement_rate_ggplot(data)
}
\description{
ggplot2: Placement rate by IQ group
}
