% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{plot_iq_internship_placement_base}
\alias{plot_iq_internship_placement_base}
\title{Base R: Grouped boxplot (IQ × Internship × Placement)}
\usage{
plot_iq_internship_placement_base(data)
}
\description{
Base R: Grouped boxplot (IQ × Internship × Placement)
}
