% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ggplot.R
\name{plot_iq_internship_placement_ggplot}
\alias{plot_iq_internship_placement_ggplot}
\title{ggplot2: Grouped boxplot (IQ × Internship × Placement)}
\usage{
plot_iq_internship_placement_ggplot(data)
}
\description{
ggplot2: Grouped boxplot (IQ × Internship × Placement)
}
