% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_placement_summary.R
\name{print.placement_summary}
\alias{print.placement_summary}
\title{Print method for Student job placement_summary objects}
\usage{
\method{print}{placement_summary}(x, ...)
}
\arguments{
\item{x}{A placement_summary object}

\item{...}{Additional arguments}
}
\description{
Print method for Student job placement_summary objects
}
