% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_placement.R
\name{summarize_placement}
\alias{summarize_placement}
\title{Summarize Student Job Placement Data}
\usage{
summarize_placement(data)
}
\arguments{
\item{data}{A cleaned placement dataset}
}
\value{
An object of class placement_summary
}
\description{
Summarize Student Job Placement Data
}
